Add files via upload
this is a Library management system project allows users to see all list of books available in library, Lending a book from library, add a book into library and search books available in library as well as display the number of books available in the library. it is an innovative system.